backend/awsscripts/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
boto3==1.35.10
botocore==1.35.99
certifi==2025.8.3
charset-normalizer==3.4.3
ecdsa==0.19.1
fastapi==0.117.1
filelock==3.19.1
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
Mako==1.3.10
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.3
rsa==4.9.1
s3transfer==0.10.4
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.3.1
torchvision==0.18.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
